requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==3.0
django-filter==23.2
djangorestframework==3.10.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
